apps/worker/requirements.txt
frictionless[csv,excel,remote]==5.17.0
requests==2.31.0
psycopg2-binary==2.9.9
redis==5.0.1
pandas==2.2.0
openpyxl==3.1.2
chardet==5.2.0
python-dotenv==1.0.1
schedule==1.2.1
